+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/Throwables.kt
@@ -1,9 +1,5 @@
 package at.asitplus.signum.supreme
 
-@RequiresOptIn(message = "Access to potentially hazardous platform-specific internals requires explicit opt-in. Specify @OptIn(HazardousMaterials::class). These accessors are unstable and may change without warning.")
-/** This is an internal property. It is exposed if you know what you are doing. You very likely don't actually need it. */
-annotation class HazardousMaterials
-
 
 @RequiresOptIn(message = "Access to secret and private key material requires explicit opt-in. Specify @OptIn(SecretExposure::class). Make sure that you actually want to externalise a secret. Check yourself, before you really, really, wreck yourself!")
 /** This guards a secret. Do not expose it lightly! */